Move kotlin lifecycle tests to AutoDisposeLifecycleTest
