+++ autodispose-kotlin/src/test/kotlin/com/uber/autodispose/kotlin/AutoDisposeKotlinTest.kt
@@ -16,9 +16,6 @@
 
 package com.uber.autodispose.kotlin
 
-import com.uber.autodispose.lifecycle.LifecycleEndedException
-import com.uber.autodispose.lifecycle.LifecycleNotStartedException
-import com.uber.autodispose.lifecycle.TestLifecycleScopeProvider
 import com.uber.autodispose.TestScopeProvider
 import io.reactivex.BackpressureStrategy.ERROR
 import io.reactivex.Completable
@@ -37,13 +34,13 @@ import org.junit.Test
 class AutoDisposeKotlinTest {
 
   companion object {
-      private const val DEFAULT_PARALLELISM = 2
+    private const val DEFAULT_PARALLELISM = 2
   }
+
   private val o = TestObserver<String>()
   private val s = TestSubscriber<String>()
   private val scopeMaybe = MaybeSubject.create<Any>()
   private val scopeProvider = TestScopeProvider.create()
-  private val lifecycleScopeProvider = com.uber.autodispose.lifecycle.TestLifecycleScopeProvider.create()
 
   @Test fun observable_maybeNormalCompletion() {
     Observable.just("Hello")
@@ -95,52 +92,6 @@ class AutoDisposeKotlinTest {
     // https://github.com/ReactiveX/RxJava/issues/5178
 //    assertThat(o.isDisposed).isTrue()
 //    o.assertNotSubscribed()
-  }
-
-  @Test fun observable_lifecycleNotStarted() {
-    Observable.just("Hello")
-        .autoDisposable(lifecycleScopeProvider)
-        .subscribe(o)
-
-    o.assertError { it is com.uber.autodispose.lifecycle.LifecycleNotStartedException }
-  }
-
-  @Test fun observable_lifecycleNormalCompletion() {
-    lifecycleScopeProvider.start()
-    Observable.just("Hello")
-        .autoDisposable(lifecycleScopeProvider)
-        .subscribe(o)
-
-    o.assertValue { it == "Hello" }
-    o.assertComplete()
-  }
-
-  @Test fun observable_lifecycleNormalInterrupted() {
-    lifecycleScopeProvider.start()
-    val subject = PublishSubject.create<String>()
-    subject
-        .autoDisposable(lifecycleScopeProvider)
-        .subscribe(o)
-
-    subject.onNext("Hello")
-
-    o.assertValue { it == "Hello" }
-
-    lifecycleScopeProvider.stop()
-
-    // https://github.com/ReactiveX/RxJava/issues/5178
-//    assertThat(o.isDisposed).isTrue()
-//    o.assertNotSubscribed()
-  }
-
-  @Test fun observable_lifecycleEnded() {
-    lifecycleScopeProvider.start()
-    lifecycleScopeProvider.stop()
-    Observable.just("Hello")
-        .autoDisposable(lifecycleScopeProvider)
-        .subscribe(o)
-
-    o.assertError { it is com.uber.autodispose.lifecycle.LifecycleEndedException }
   }
 
   @Test fun flowable_maybeNormalCompletion() {
@@ -195,52 +146,6 @@ class AutoDisposeKotlinTest {
 //    s.assertNotSubscribed()
   }
 
-  @Test fun flowable_lifecycleNotStarted() {
-    Flowable.just("Hello")
-        .autoDisposable(lifecycleScopeProvider)
-        .subscribe(s)
-
-    s.assertError { it is com.uber.autodispose.lifecycle.LifecycleNotStartedException }
-  }
-
-  @Test fun flowable_lifecycleNormalCompletion() {
-    lifecycleScopeProvider.start()
-    Flowable.just("Hello")
-        .autoDisposable(lifecycleScopeProvider)
-        .subscribe(s)
-
-    s.assertValue { it == "Hello" }
-    s.assertComplete()
-  }
-
-  @Test fun flowable_lifecycleNormalInterrupted() {
-    lifecycleScopeProvider.start()
-    val subject = PublishSubject.create<String>()
-    subject.toFlowable(ERROR)
-        .autoDisposable(lifecycleScopeProvider)
-        .subscribe(s)
-
-    subject.onNext("Hello")
-
-    s.assertValue { it == "Hello" }
-
-    lifecycleScopeProvider.stop()
-
-    // https://github.com/ReactiveX/RxJava/issues/5178
-//    assertThat(s.isDisposed).isTrue()
-//    s.assertNotSubscribed()
-  }
-
-  @Test fun flowable_lifecycleEnded() {
-    lifecycleScopeProvider.start()
-    lifecycleScopeProvider.stop()
-    Flowable.just("Hello")
-        .autoDisposable(lifecycleScopeProvider)
-        .subscribe(s)
-
-    s.assertError { it is com.uber.autodispose.lifecycle.LifecycleEndedException }
-  }
-
   @Test fun maybe_maybeNormalCompletion() {
     Maybe.just("Hello")
         .autoDisposable(scopeMaybe)
@@ -291,48 +196,6 @@ class AutoDisposeKotlinTest {
     // https://github.com/ReactiveX/RxJava/issues/5178
 //    assertThat(o.isDisposed).isTrue()
 //    o.assertNotSubscribed()
-  }
-
-  @Test fun maybe_lifecycleNotStarted() {
-    Maybe.just("Hello")
-        .autoDisposable(lifecycleScopeProvider)
-        .subscribe(o)
-
-    o.assertError { it is com.uber.autodispose.lifecycle.LifecycleNotStartedException }
-  }
-
-  @Test fun maybe_lifecycleNormalCompletion() {
-    lifecycleScopeProvider.start()
-    Maybe.just("Hello")
-        .autoDisposable(lifecycleScopeProvider)
-        .subscribe(o)
-
-    o.assertValue { it == "Hello" }
-    o.assertComplete()
-  }
-
-  @Test fun maybe_lifecycleNormalInterrupted() {
-    lifecycleScopeProvider.start()
-    val subject = PublishSubject.create<String>()
-    subject
-        .autoDisposable(lifecycleScopeProvider)
-        .subscribe(o)
-
-    lifecycleScopeProvider.stop()
-
-    // https://github.com/ReactiveX/RxJava/issues/5178
-//    assertThat(o.isDisposed).isTrue()
-//    o.assertNotSubscribed()
-  }
-
-  @Test fun maybe_lifecycleEnded() {
-    lifecycleScopeProvider.start()
-    lifecycleScopeProvider.stop()
-    Maybe.just("Hello")
-        .autoDisposable(lifecycleScopeProvider)
-        .subscribe(o)
-
-    o.assertError { it is com.uber.autodispose.lifecycle.LifecycleEndedException }
   }
 
   @Test fun single_maybeNormalCompletion() {
@@ -387,48 +250,6 @@ class AutoDisposeKotlinTest {
 //    o.assertNotSubscribed()
   }
 
-  @Test fun single_lifecycleNotStarted() {
-    Single.just("Hello")
-        .autoDisposable(lifecycleScopeProvider)
-        .subscribe(o)
-
-    o.assertError { it is com.uber.autodispose.lifecycle.LifecycleNotStartedException }
-  }
-
-  @Test fun single_lifecycleNormalCompletion() {
-    lifecycleScopeProvider.start()
-    Single.just("Hello")
-        .autoDisposable(lifecycleScopeProvider)
-        .subscribe(o)
-
-    o.assertValue { it == "Hello" }
-    o.assertComplete()
-  }
-
-  @Test fun single_lifecycleNormalInterrupted() {
-    lifecycleScopeProvider.start()
-    val subject = PublishSubject.create<String>()
-    subject
-        .autoDisposable(lifecycleScopeProvider)
-        .subscribe(o)
-
-    lifecycleScopeProvider.stop()
-
-    // https://github.com/ReactiveX/RxJava/issues/5178
-//    assertThat(o.isDisposed).isTrue()
-//    o.assertNotSubscribed()
-  }
-
-  @Test fun single_lifecycleEnded() {
-    lifecycleScopeProvider.start()
-    lifecycleScopeProvider.stop()
-    Single.just("Hello")
-        .autoDisposable(lifecycleScopeProvider)
-        .subscribe(o)
-
-    o.assertError { it is com.uber.autodispose.lifecycle.LifecycleEndedException }
-  }
-
   @Test fun completable_maybeNormalCompletion() {
     Completable.complete()
         .autoDisposable(scopeMaybe)
@@ -477,50 +298,9 @@ class AutoDisposeKotlinTest {
 //    o.assertNotSubscribed()
   }
 
-  @Test fun completable_lifecycleNotStarted() {
-    Completable.complete()
-        .autoDisposable(lifecycleScopeProvider)
-        .subscribe(o)
-
-    o.assertError { it is com.uber.autodispose.lifecycle.LifecycleNotStartedException }
-  }
-
-  @Test fun completable_lifecycleNormalCompletion() {
-    lifecycleScopeProvider.start()
-    Completable.complete()
-        .autoDisposable(lifecycleScopeProvider)
-        .subscribe(o)
-
-    o.assertComplete()
-  }
-
-  @Test fun completable_lifecycleNormalInterrupted() {
-    lifecycleScopeProvider.start()
-    val subject = CompletableSubject.create()
-    subject
-        .autoDisposable(lifecycleScopeProvider)
-        .subscribe(o)
-
-    lifecycleScopeProvider.stop()
-
-    // https://github.com/ReactiveX/RxJava/issues/5178
-//    assertThat(o.isDisposed).isTrue()
-//    o.assertNotSubscribed()
-  }
-
-  @Test fun completable_lifecycleEnded() {
-    lifecycleScopeProvider.start()
-    lifecycleScopeProvider.stop()
-    Completable.complete()
-        .autoDisposable(lifecycleScopeProvider)
-        .subscribe(o)
-
-    o.assertError { it is com.uber.autodispose.lifecycle.LifecycleEndedException }
-  }
-
   @Test fun parallelFlowable_maybeNormalCompletion() {
     val s2 = TestSubscriber<String>()
-    Flowable.just(  "Hello", "World")
+    Flowable.just("Hello", "World")
         .parallel(DEFAULT_PARALLELISM)
         .autoDisposable(scopeMaybe)
         .subscribe(arrayOf(s, s2))
@@ -543,7 +323,7 @@ class AutoDisposeKotlinTest {
     subject.onNext("World")
 
     s.assertValue { it == "Hello" }
-    s2.assertValue{ it == "World" }
+    s2.assertValue { it == "World" }
 
     scopeMaybe.onSuccess(Object())
 
@@ -584,65 +364,6 @@ class AutoDisposeKotlinTest {
 // https://github.com/ReactiveX/RxJava/issues/5178
 //    assertThat(s.isDisposed).isTrue()
 //    s.assertNotSubscribed()
-  }
-
-  @Test fun parallelFlowable_lifecycleNotStarted() {
-    val s2 = TestSubscriber<String>()
-    Flowable.just("Hello", "World")
-        .parallel(DEFAULT_PARALLELISM)
-        .autoDisposable(lifecycleScopeProvider)
-        .subscribe(arrayOf(s, s2))
-
-    s.assertError { it is com.uber.autodispose.lifecycle.LifecycleNotStartedException }
-    s2.assertError { it is com.uber.autodispose.lifecycle.LifecycleNotStartedException }
-  }
-
-  @Test fun parallelFlowable_lifecycleNormalCompletion() {
-    lifecycleScopeProvider.start()
-    val s2 = TestSubscriber<String>()
-    Flowable.just("Hello", "World")
-        .parallel(DEFAULT_PARALLELISM)
-        .autoDisposable(lifecycleScopeProvider)
-        .subscribe(arrayOf(s, s2))
-
-    s.assertValue { it == "Hello" }
-    s.assertComplete()
-    s2.assertValue { it == "World" }
-    s2.assertComplete()
-  }
-
-  @Test fun parallelFlowable_lifecycleNormalInterrupted() {
-    lifecycleScopeProvider.start()
-    val subject = PublishSubject.create<String>()
-    val s2 = TestSubscriber<String>()
-    subject.toFlowable(ERROR)
-        .parallel(DEFAULT_PARALLELISM)
-        .autoDisposable(lifecycleScopeProvider)
-        .subscribe(arrayOf(s, s2))
-
-    subject.onNext("Hello")
-    subject.onNext("World")
-
-    s.assertValue { it == "Hello" }
-    s2.assertValue { it == "World" }
-    lifecycleScopeProvider.stop()
-
-// https://github.com/ReactiveX/RxJava/issues/5178
-//    assertThat(s.isDisposed).isTrue()
-//    s.assertNotSubscribed()
-  }
-
-  @Test fun parallelFlowable_lifecycleEnded() {
-    lifecycleScopeProvider.start()
-    lifecycleScopeProvider.stop()
-    val s2 = TestSubscriber<String>()
-    Flowable.just("Hello")
-        .parallel(DEFAULT_PARALLELISM)
-        .autoDisposable(lifecycleScopeProvider)
-        .subscribe(arrayOf(s, s2))
-
-    s.assertError { it is com.uber.autodispose.lifecycle.LifecycleEndedException }
-    s2.assertError { it is com.uber.autodispose.lifecycle.LifecycleEndedException }
   }
 
 }
